--- src/css/Admin/AddEmployee.css
@@ -1,68 +1,0 @@
-/* AddEmployee.css */
-
-.modal-content {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.user-list {
-  margin-top: 20px;
-  width: 100%;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-th, td {
-  border: 1px solid #dddddd;
-  padding: 8px;
-  text-align: left;
-}
-
-th {
-  background-color: #f2f2f2;
-}
-
-.button-container {
-  margin-top: 20px;
-  align-self: flex-end;
-}
-
-.add-button {
-  padding: 10px 20px;
-  margin-right: 10px;
-  background-color: #4CAF50;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  text-align: center;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 16px;
-  transition-duration: 0.4s;
-}
-.remove-button{
-  padding: 10px 20px;
-  margin-right: 10px;
-  background-color: #b61d18;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  text-align: center;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 16px;
-  transition-duration: 0.4s;
-}
-
-.add-button:hover{
-  background-color: #45a049;
-}
-.remove-button:hover{
-    background-color: #af3430;
-}
