+++ src/css/Admin/ManageUsers.css
@@ -1,0 +1,113 @@
+/* ManageUsers.css */
+
+.manage-users-modal {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: rgba(0, 0, 0, 0.5); /* Semi-transparent background */
+    z-index: 999; /* Ensure the modal is on top */
+  }
+  
+  .manage-users-modal .modal-content {
+    background-color: #fff;
+    width: 80%;
+    max-width: 950px;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.2); /* Add a subtle shadow */
+  }
+  
+  .manage-users-modal .modal-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 20px;
+  }
+  
+  .manage-users-modal .modal-header h1 {
+    margin: 0;
+  }
+  
+  .manage-users-modal .close {
+    cursor: pointer;
+  }
+  
+  .manage-users-modal .user-list {
+    margin-top: 20px;
+  }
+  
+  .manage-users-modal table {
+    width: 100%;
+    border-collapse: collapse;
+  }
+  
+  .manage-users-modal th,
+  .manage-users-modal td {
+    border: 1px solid #dddddd;
+    padding: 8px;
+    text-align: left;
+  }
+  
+  .manage-users-modal th {
+    background-color: #f2f2f2;
+  }
+  
+  .manage-users-modal .button-container {
+    margin-top: 20px;
+    text-align: right;
+  }
+  
+  .manage-users-modal button {
+    padding: 10px 20px;
+    margin: 5px;
+    background-color: #4CAF50;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 16px;
+    transition-duration: 0.4s;
+  }
+  
+
+  
+  /* Specific button styles */
+  .manage-users-modal .add-employee-button {
+    background-color: #4CAF50; /* Green */
+  }
+
+  .manage-users-modal .add-employee-button:hover {
+    background-color: #7cb17d; /* Green */
+  }
+  
+  
+  .manage-users-modal .remove-employee-button {
+    background-color: #c92a2a; /* Light Red */
+  }
+  .manage-users-modal .remove-employee-button:hover {
+    background-color: #f05656; /* Light Red */
+  }
+  
+  .manage-users-modal .ban-button {
+    background-color: #8B0000; /* Dark Red */
+  }
+  .manage-users-modal .ban-button:hover {
+    background-color: #ff0606; /* Dark Red */
+  }
+  
+  .manage-users-modal .view-order-history-button {
+    background-color: #ff5100; /* Orange */
+  }
+  .manage-users-modal .view-order-history-button:hover {
+    background-color: #f08e60; /* Orange */
+  }
+
+
